+++ src/api/test_env.py
@@ -2,7 +2,7 @@ import os
 
 import pytest
 
-from . import credentials as C
+from . import environments as E
 
 
 @pytest.mark.parametrize(
@@ -16,7 +16,7 @@ def test_invalid_env_credentials(monkeypatch, env_attr):
         monkeypatch.delenv(env_attr)
 
     with pytest.raises(KeyError):
-        C.env_credentials()
+        E.env_credentials()
 
 
 def test_invalid_env_region(monkeypatch):
@@ -26,9 +26,9 @@ def test_invalid_env_region(monkeypatch):
         monkeypatch.delenv(env_attr)
 
     with pytest.raises(KeyError):
-        C.env_region()
+        E.env_region()
 
     monkeypatch.setenv(env_attr, "NaN")
     with pytest.raises(ValueError):
-        C.env_region()
+        E.env_region()
